requirements.txt
Flask==0.10.1
gunicorn==19.1.1
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.7.3
Werkzeug==0.9.6
wsgiref==0.1.2
MarkupSafe==0.23
psycopg2==2.5.4
requests
grequests
lxml
pickle
argparse
unittest
urllib2
logging
